requirements.txt
Flask
Flask-RESTful
Flask-JWT
mysql-connector-python
beautifulsoup4
requests
uwsgi
psycopg2